Joined ide and help to one tab && added ide options
