+++ src/components/IDEHelpPanel/index.js
@@ -1,0 +1,3 @@
+import IDEHelpPanel from './IDEHelpPanel'
+
+export default IDEHelpPanel
