+++ src/components/IDEOptionsPanel/index.js
@@ -1,0 +1,3 @@
+import IDEOptionsPanel from './IDEOptionsPanel'
+
+export default IDEOptionsPanel
