+++ src/components/IDEPanel/index.js
@@ -1,0 +1,3 @@
+import IDEPanel from './IDEPanel'
+
+export default IDEPanel
