+++ src/components/StrategyEditor/style.scss
@@ -270,7 +270,7 @@ body.hosted {
     }
   }
   .title-label {
-    margin-right: 10px;
+    margin-right: 5px;
     &:last-child {
       margin-right: 0;
     }
