+++ src/components/StrategyEditor/tabs/IDETab.Title.js
@@ -1,21 +1,13 @@
-import React, { memo, useMemo } from 'react'
+import React, { memo } from 'react'
 import PropTypes from 'prop-types'
-import _values from 'lodash/values'
-import _some from 'lodash/some'
 import { useTranslation } from 'react-i18next'
 import { Icon } from 'react-fa'
 import Indicator from '../../../ui/Indicator'
 
-const IDETabTitle = ({ sidebarOpened, strategyDirty, sectionErrors }) => {
+const IDETabTitle = ({
+  sidebarOpened, strategyDirty, hasErrors,
+}) => {
   const { t } = useTranslation()
-
-  const hasErrors = useMemo(() => {
-    if (!sectionErrors) {
-      return false
-    }
-    const errors = _values(sectionErrors)
-    return _some(errors, (error) => error)
-  }, [sectionErrors])
 
   const getIndicator = () => {
     if (hasErrors) {
@@ -39,11 +31,7 @@ const IDETabTitle = ({ sidebarOpened, strategyDirty, sectionErrors }) => {
 IDETabTitle.propTypes = {
   sidebarOpened: PropTypes.bool.isRequired,
   strategyDirty: PropTypes.bool.isRequired,
-  sectionErrors: PropTypes.objectOf(PropTypes.string),
-}
-
-IDETabTitle.defaultProps = {
-  sectionErrors: null,
+  hasErrors: PropTypes.bool.isRequired,
 }
 
 export default memo(IDETabTitle)
