+++ src/ui/Markdown/style.scss
@@ -41,7 +41,6 @@
   h2 { 
     border-bottom: 1px solid gray;
     padding-bottom: 10px;
-    margin-top: 20px;
     color: $code-anchor-color;
     code { 
       color: $code-text-color;
