+++ src/ui/Panel/style.scss
@@ -309,7 +309,7 @@
 
     &.sidebar-opened {
       .hfui_panel__sidebar-container {
-        width: 150px;
+        width: 155px;
       }
     }
